requirements.txt
nltk
python-twitter
